refactor(utils): consolidate shared utilities from web (#408)
Migrate validate-month, sorting, and get-unique-months to @sgcarstrends/utils.
Eliminate apps/web/src/utils/arrays/ directory by flattening group-by-year.
