--- apps/web/src/utils/arrays/index.ts
@@ -1,3 +1,0 @@
-export { getUniqueMonths } from "./get-unique-months";
-export { groupByYear } from "./group-by-year";
-export { SortDirection, sortByName, sortByValue } from "./sorting";
