+++ packages/utils/src/index.ts
@@ -6,6 +6,9 @@ export * from "./format-date-to-month-year";
 export * from "./format-growth-rate";
 export * from "./format-ordinal";
 export * from "./format-percentage";
+export * from "./get-unique-months";
 export * from "./redis";
 export * from "./slugify";
+export * from "./sorting";
 export * from "./tokeniser";
+export * from "./validate-month";
